Implementing file uploading to dropbox and similar stuff.

+++ SaitynoProjektasBackEnd/SaitynoProjektasBackEnd/Services/SongsService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using SaitynoProjektasBackEnd.Data;
 using SaitynoProjektasBackEnd.Models;
@@ -12,10 +13,12 @@ namespace SaitynoProjektasBackEnd.Services
     public class SongsService : ISongsService
     {
         private readonly ApplicationDbContext _context;
+        private readonly IDropBoxService _dropBoxService;
 
-        public SongsService(ApplicationDbContext context)
+        public SongsService(ApplicationDbContext context, IDropBoxService dropBoxService)
         {
             _context = context;
+            _dropBoxService = dropBoxService;
         }
 
         public IEnumerable<SongResponseModel> GetSongs()
@@ -70,17 +73,27 @@ namespace SaitynoProjektasBackEnd.Services
             return songResponseModel;
         }
 
-        public string[] AddSong(AddSongRequestModel songRequestModel, string userName)
+        public async Task<string[]> AddSong(AddSongRequestModel songRequestModel, string userName)
         {
             var genre = _context.Genres
                 .SingleOrDefault(g => g.Name == songRequestModel.Genre);
             var user = _context.Users
                 .SingleOrDefault(u => u.UserName == userName);
+            var isFilePathUsed = _context.Songs
+                .Where(s => s.FilePath == songRequestModel.FilePath)
+                .Any();
 
             if (genre == null)
                 return new[] {"Genre is not found"};
             if (user == null)
                 return new[] {"User is not found"};
+            if (isFilePathUsed)
+                return new[] {"Invalid file path"};
+
+            var fileIsFound = await _dropBoxService.DoesFileExist(songRequestModel.FilePath);
+
+            if (!fileIsFound)
+                return new[] {"File specified is not found in the file storage"};
 
             var song = new Song
             {
@@ -91,7 +104,8 @@ namespace SaitynoProjektasBackEnd.Services
                 Duration = 0,
                 Plays = 0,
                 Genre = genre,
-                User = user
+                User = user,
+                FilePath = songRequestModel.FilePath
             };
 
             _context.Songs.Add(song);
@@ -100,15 +114,20 @@ namespace SaitynoProjektasBackEnd.Services
             return result == 0 ? new[] {"Could not add song"} : null;
         }
 
-        public string[] EditSong(int id, EditSongRequestModel songRequestModel)
+        public string[] EditSong(int id, EditSongRequestModel songRequestModel, string userName)
         {
             var song = _context.Songs
+                .Include(s => s.User)
                 .SingleOrDefault(s => s.Id == id);
+            var user = _context.Users
+                .SingleOrDefault(u => u.UserName == userName);
 
             if (song == null)
-            {
                 return new[] {"Song is not found"};
-            }
+            if (user == null)
+                return new[] {"User is not found"};
+            if (song.User.UserName != userName)
+                return new[] {"User is not the owner of the song"};
 
             if (!string.IsNullOrEmpty(songRequestModel.Title))
             {
@@ -125,15 +144,20 @@ namespace SaitynoProjektasBackEnd.Services
             return null;
         }
 
-        public string[] DeleteSong(int id)
+        public async Task<string[]> DeleteSong(int id, string userName)
         {
             var song = _context.Songs
+                .Include(s => s.User)
                 .SingleOrDefault(s => s.Id == id);
+            var user = _context.Users
+                .SingleOrDefault(u => u.UserName == userName);
 
             if (song == null)
-            {
                 return new[] {"Song is not found"};
-            }
+            if (user == null)
+                return new[] {"User is not found"};
+            if (song.User.UserName != userName)
+                return new[] {"User is not the owner of the song"};
 
             var playlistSongs = _context.PlaylistSongs.Where(ps => ps.SongId == id);
             _context.PlaylistSongs.RemoveRange(playlistSongs);
@@ -144,6 +168,8 @@ namespace SaitynoProjektasBackEnd.Services
             _context.Songs.Remove(song);
 
             _context.SaveChanges();
+
+            var result = await _dropBoxService.DeleteFile(song.FilePath);
 
             return null;
         }
